+++ app/Http/Controllers/Admin/StoreController.php
@@ -5,14 +5,22 @@ namespace App\Http\Controllers\Admin;
 use App\User;
 use App\Store;
 use App\Http\Requests\StoreRequest;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Traits\UploadTrait;
+use Illuminate\Support\Facades\Storage;
 
 class StoreController extends Controller
 {
+
+    use UploadTrait;
+
+    public function __construct()
+    {
+        $this->middleware('user.has.store')->only(['create', 'store']);
+    }
+
     public function index()
     {
-        // $stores = Store::paginate(10);
 
         $store = auth()->user()->store;
 
@@ -21,11 +29,6 @@ class StoreController extends Controller
 
     public function create()
     {
-        if(auth()->user()->store()->count()){
-            flash('Você já possui uma Loja!')->warning();
-            return redirect()->route('admin.stores.index');
-        }
-
         $users = User::all(['id', 'name']);
 
         return view('admin.stores.create', compact('users'));
@@ -37,11 +40,15 @@ class StoreController extends Controller
 
         $user = auth()->user();
 
-        $user->store()->create($data);
+        if($request->hasFile('logo')){
+            $data['logo'] = $this->imageUpload($request->file('logo'));
+        }
+
+        $store = $user->store()->create($data);
 
         flash('Loja Criada com Sucesso!')->success();
 
-        return redirect()->route('admin.stores.index');
+        return redirect()->route('admin.stores.edit', ['store' => $store->id]);
     }
 
     public function edit($store)
@@ -56,21 +63,35 @@ class StoreController extends Controller
         $data = $request->all();
 
         $store = Store::find($store);
+
+        if($request->hasFile('logo')){
+            if(Storage::disk('public')->exists($store->logo)){
+                Storage::disk('public')->delete($store->logo);
+            }
+                $data['logo'] = $this->imageUpload($request->file('logo'));
+        }
+
         $store->update($data);
 
         flash('Loja Atualizada com Sucesso!')->success();
 
-        return redirect()->route('admin.stores.index');
+        return redirect()->route('admin.stores.edit', ['store' => $store->id]);
     }
 
     public function destroy($store)
     {
+        $validStore = auth()->user()->store;
+        if($validStore->products()->count()){
+            flash('Você possui produtos relacionados a esta loja!')->warning();
+            return redirect()->route('admin.stores.index');
+        }
+
         $store = Store::find($store);
 
         $store->delete();
 
         flash('Loja Removida com Sucesso!')->success();
 
-        return redirect()->route('admin.stores.index');
+        return redirect()->back();
     }
 }
